transcript-generation/pyproject.toml
[project]
name = "transcript-generation"
version = "0.1.0"
description = "Poll recent videos from House Committee YouTube channels"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.1.0",
    "rich>=13.0.0",
    "yt-dlp>=2025.9.5",
    "openai>=1.40.0",
    "assemblyai>=0.43.1",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
]
